requirements.txt
tensorflow
flask
opencv-python
matplotlib
scikit-learn
numpy
Pillow
